requirements.txt
altgraph==0.17.4
arrow==1.3.0
asttokens==2.4.1
beautifulsoup4==4.12.3
binaryornot==0.4.4
blinker==1.8.2
botcity-framework-base==0.4.4
botcity-framework-core==1.2.1
botcity-framework-web==0.9.0
botcity-http-plugin==0.1.4
botcity-maestro-sdk==0.5.1
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
comtypes==1.4.7
cookiecutter==2.6.0
decorator==5.1.1
distro==1.9.0
et-xmlfile==1.1.0
executing==2.1.0
Flask==3.0.3
idna==3.10
importlib_metadata==8.5.0
ipython==8.28.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==3.0.1
matplotlib-inline==0.1.7
mdurl==0.1.2
msedge-selenium-tools==3.141.3
mysql-connector-python==9.1.0
numpy==1.26.4
opencv-python==4.10.0.84
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
parso==0.8.4
pefile==2023.2.7
pillow==10.4.0
prompt_toolkit==3.0.48
psutil==6.0.0
pure_eval==0.2.3
Pygments==2.18.0
pyinstaller==6.11.0
pyinstaller-hooks-contrib==2024.9
pynput==1.7.7
pyperclip==1.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
python-xlib==0.33
pytz==2024.2
pywin32==308
pywin32-ctypes==0.2.3
pywinauto==0.6.8
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.2
selenium==3.141.0
setuptools==75.2.0
six==1.16.0
soupsieve==2.6
stack-data==0.6.3
text-unidecode==1.3
traitlets==5.14.3
types-python-dateutil==2.9.0.20241003
tzdata==2024.2
urllib3==1.26.20
wcwidth==0.2.13
webdriver-manager==4.0.2
Werkzeug==3.1.2
zipp==3.20.2
